+++ apps/www/src/routes/icons.tsx
@@ -34,7 +34,6 @@ const names: IconName[] = [
   "close",
   "copy",
   "credit-card",
-  "github",
   "info",
   "italic",
   "loader",